JL8-27 fix : 가사 라우트 핸들러로 수정
