+++ src/app/api/lyrics/token/route.ts
