+++ src/hooks/usePlayer.ts
@@ -1,9 +1,8 @@
 'use client'
 import { fetchMusicDetailByMusicId } from '@/api/music-play/actions'
-import { getSongLyrics } from '@/api/music-play/genius-api'
 import { fetchPreviewUrl } from '@/api/spotifyToken'
 import { useMusicPlayerStore } from '@/store/useMusicPlayerStore'
-import { getAccessToken } from '@/utils/geniusToken'
+import { getSongLyrics } from '@/utils/geniusApi'
 import { useQuery } from '@tanstack/react-query'
 
 const usePlayer = () => {
@@ -14,19 +13,17 @@ const usePlayer = () => {
   const { data: musicDetail, isPending } = useQuery({
     queryKey: ['music', currentTrackId],
     queryFn: async () => {
-      const [accessToken, trackUrl, musicDetail] = await Promise.all([
-        getAccessToken(),
+      const [trackUrl, musicDetail] = await Promise.all([
         fetchPreviewUrl(currentTrackId),
         fetchMusicDetailByMusicId(currentTrackId),
       ])
 
       // 노래 가사 가져오기
       let lyrics = null
-      if (musicDetail && accessToken) {
+      if (musicDetail) {
         lyrics = await getSongLyrics({
           title: musicDetail.title,
           artist: musicDetail.artist,
-          accessToken,
         })
       }
